Fix potential overflow in ogg_stream_iovecin.
The original commit from 2007 (440e37e72223) accumulated all sizes
 in ints (despite ogg_iovec_t.iov_len being a size_t and all
 relevant fields of ogg_stream_state being longs).
I expanded this to longs in 2013 when I added overflow checks
 (commit 85dbd8d9ed3d), but apparently did not read far enough down
 in the function to catch this cast back to int the second time the
 sizes are accumulated.

Patch from: Chengyu Song <csong@ucr.edu>

Fixes #2308.

+++ src/framing.c
@@ -312,7 +312,7 @@ int ogg_stream_iovecin(ogg_stream_state *os, ogg_iovec_t *iov, int count,
 
   for (i = 0; i < count; ++i) {
     memcpy(os->body_data+os->body_fill, iov[i].iov_base, iov[i].iov_len);
-    os->body_fill += (int)iov[i].iov_len;
+    os->body_fill += (long)iov[i].iov_len;
   }
 
   /* Store lacing vals for this packet */
